requirements.txt
opencv-python
torch
torchvision
transformers
numpy
pandas
matplotlib
scikit-learn
Pillow
tqdm